chore: Move emails into their own directory
