+++ src/controllers/email-preview.js
@@ -7,13 +7,13 @@ import AppConstants from '../app-constants.js'
 import { notify } from './hibp.js'
 import { mainLayout } from '../views/main.js'
 import { emailPreview } from '../views/partials/email-preview.js'
-import { getTemplate, getPreviewTemplate } from '../views/email-2022.js'
-import { breachAlertEmailPartial } from '../views/partials/email-breach-alert.js'
-import { signupReportEmailPartial } from '../views/partials/email-signup-report.js'
-import { verifyPartial } from '../views/partials/email-verify.js'
+import { getTemplate, getPreviewTemplate } from '../views/emails/email-2022.js'
+import { breachAlertEmailPartial } from '../views/emails/email-breach-alert.js'
+import { signupReportEmailPartial } from '../views/emails/email-signup-report.js'
+import { verifyPartial } from '../views/emails/email-verify.js'
 import {
   monthlyUnresolvedEmailPartial
-} from '../views/partials/email-monthly-unresolved.js'
+} from '../views/emails/email-monthly-unresolved.js'
 
 import { getMessage } from '../utils/fluent.js'
 import { generateToken } from '../utils/csrf.js'
