+++ src/views/emails/email-2022.js
@@ -2,8 +2,8 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-import AppConstants from '../app-constants.js'
-import { getMessage } from '../utils/fluent.js'
+import AppConstants from '../../app-constants.js'
+import { getMessage } from '../../utils/fluent.js'
 
 const companyAddress = '2 Harrison St. #175, San Francisco, California 94105 USA'
 const links = (data) => ({
