+++ src/views/emails/email-signup-report.js
