+++ src/views/emails/email-verify.js
